testrunner latencyBudget test fixed Problem was that curPos was a uint8_t, but buffSize was 1024.

+++ include/sdds/NetBuffRef.h
@@ -29,7 +29,7 @@ struct _NetBuffRef_t {
 
     NetFrameBuff frame_start;
     byte_t* buff_start;
-    uint8_t curPos;
+    uint16_t curPos;
 
     // QOS part
 #if SDDS_QOS_DW_LARBUD < 65536
